pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ragversion"
version = "0.2.0"
description = "Async-first version tracking system for RAG applications"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "Sourangshu Pal", email = "paulbindass@gmail.com" }
]
keywords = ["rag", "version-control", "document-tracking", "langchain", "llamaindex"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aiofiles>=23.0.0",
    "python-magic>=0.4.27",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
parsers = [
    "pypdf>=3.0.0",
    "python-docx>=1.0.0",
    "python-pptx>=0.6.21",
    "openpyxl>=3.1.0",
    "markdown>=3.5.0",
]
langchain = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.1",
]
llamaindex = [
    "llama-index>=0.9.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]
all = [
    "ragversion[parsers,langchain,llamaindex,dev]",
]

[project.scripts]
ragversion = "ragversion.cli:main"

[project.urls]
Homepage = "https://github.com/sourangshupal/ragversion"
Repository = "https://github.com/sourangshupal/ragversion"
Issues = "https://github.com/sourangshupal/ragversion/issues"

[tool.hatch.build.targets.wheel]
packages = ["ragversion"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--cov=ragversion",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "docx.*",
    "pptx.*",
    "magic.*",
    "supabase.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["ragversion"]
omit = ["*/tests/*", "*/testing/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
